Header accessibility updates (#7)
* Update language attribute on language switcher links as per Trello ticket #328

There are some links for switching language. This commit adds on the lang attribute with the appropriate value, so that a screen reader will understand better how to pronounce them.

* Fix menu to the top of the screen with hamburger menu in it as per Trello ticket #327

The header wasn't sticking to the top, and the hamburger menu wasn't sitting inside of it either. This fixes that. It also fixes a resulting issue, where form fields were covering the navbar as you scrolled past them.