README updates (#6)
* Add in information about compiling styles to README

I had to run 'npm run watch' to get the styles to appear in the browser. Adding this in to the documentation so it's there for the next person.

* Make README more tailored to repo

There is some copy that seems to have been copied over from somewhere else, that isn't relevant to this repo. I've updated the title and repo. I've also added and removed some instructions from the 'Getting started' guide

* Add clarification around .gitkeep file

This is based on some feedback from Damien - thank you!