Add basic structure for hamburger menu accessibility improvements - Trello ticket #354
This adds in the majority of the code tweaks to make the hamburger menu more accessible. It alters the hamburger icon to a button so it's keyboard accessible, makes the menu full screen, and prevents scrolling in the background. There are a few outstanding issues which will be addressed in future commits.