Hamburger menu accessibility (#8)
* Update language attribute on language switcher links as per Trello ticket #328

There are some links for switching language. This commit adds on the lang attribute with the appropriate value, so that a screen reader will understand better how to pronounce them.

* Fix menu to the top of the screen with hamburger menu in it as per Trello ticket #327

The header wasn't sticking to the top, and the hamburger menu wasn't sitting inside of it either. This fixes that. It also fixes a resulting issue, where form fields were covering the navbar as you scrolled past them.

* Add basic structure for hamburger menu accessibility improvements - Trello ticket #354

This adds in the majority of the code tweaks to make the hamburger menu more accessible. It alters the hamburger icon to a button so it's keyboard accessible, makes the menu full screen, and prevents scrolling in the background. There are a few outstanding issues which will be addressed in future commits.

* Fixes bug where you can only close hamburger menu on touch device by swipe as per ticket #324

There was an issue with the jQuery which meant that you could only close the hamburger menu by swiping on it. Tapping on it wouldn't open it or close it. This commit fixes this bug.