Update archive page to support category returned results
Modify the archive page so that it returns a list of posts with a short
excerpt and category linked tags under those excerpts. Tweaks to search
landing page including to the search fields that determine how you
filter content. These were made after consulting the team. It was
determined that these changes were best as MVP to roll out.